refactor: Score 생성 로직을 Rule이 갖도록 변경
