+++ src/test/java/domain/rule/RuleTest.java
@@ -4,6 +4,7 @@ import domain.card.Card;
 import domain.card.Hand;
 import domain.card.Symbol;
 import domain.card.Type;
+import domain.result.Score;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,6 +12,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -79,4 +82,35 @@ class RuleTest {
 				Arguments.of(new Card(Symbol.ACE, Type.DIAMOND), false));
 	}
 
+
+	@ParameterizedTest
+	@MethodSource("calculateInput")
+	@DisplayName("스코어 계산결과가 올바른지")
+	void calculateScore(List<Card> cards, int expected) {
+		Hand hand = Hand.createEmpty();
+		for (Card card : cards) {
+			hand.add(card);
+		}
+
+		assertThat(Rule.newScore(hand)).isEqualTo(Score.from(expected));
+	}
+
+	static Stream<Arguments> calculateInput() {
+		Card ACE = new Card(Symbol.ACE, Type.CLUB);
+		Card THREE = new Card(Symbol.THREE, Type.CLUB);
+		Card SIX = new Card(Symbol.SIX, Type.CLUB);
+		Card SEVEN = new Card(Symbol.SEVEN, Type.CLUB);
+		Card EIGHT = new Card(Symbol.EIGHT, Type.CLUB);
+		Card NINE = new Card(Symbol.NINE, Type.CLUB);
+		Card TEN = new Card(Symbol.TEN, Type.CLUB);
+
+		return Stream.of(Arguments.of(Arrays.asList(EIGHT, SEVEN, SIX), 21),
+				Arguments.of(Arrays.asList(EIGHT, NINE), 17),
+				Arguments.of(Arrays.asList(ACE, ACE, ACE), 13),
+				Arguments.of(Arrays.asList(ACE, THREE), 14),
+				Arguments.of(Arrays.asList(ACE, NINE, TEN), 20),
+				Arguments.of(Arrays.asList(TEN, TEN, TEN), 30),
+				Arguments.of(Arrays.asList(ACE, TEN, TEN, ACE), 22),
+				Arguments.of(Arrays.asList(ACE, TEN, TEN), 21));
+	}
 }
